[Predator] Handle dependency rolls with old_revision missing.
If we only miss one of the 'old_revision' and 'new_revision', we replace the missing one with None. For example if we miss 'old_revision', the regression range should be (None, 'new_revision')
If we miss both 'old_revision' and 'new_revision', the regression range should be None.

Bug: 768891
Change-Id: I987f015962c5ada81b4032dee672144daeca03ad
Reviewed-on: https://chromium-review.googlesource.com/685263
Commit-Queue: Sharu Jiang <katesonia@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>