android_docker: Move mounted work dirs back to /b/.
All this broke with crrev.com/4a71824a0c1ab29de404803733fbe43ab7575052

Currently, the workdirs are not persisting after a container exits. So
every 4 hours, each container starts with an empty cache.

Also, it's not using the same work dirs it was before, which are still
present on the bots. So now we have 2x the amount of isolate caches on
the bots.

This reverts it back to the way it was before.

Bug: 768915
Change-Id: Id73fe3cb4287f83e3cb00377b7c0a7dfdbc5437d
TBR=jbudorick@chromium.org
Reviewed-on: https://chromium-review.googlesource.com/685541
Reviewed-by: Benjamin Pastene <bpastene@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Benjamin Pastene <bpastene@chromium.org>