[buildbucket] remove usage of toplevel
toplevel is a convenience function that allows handler to not
explictly wait for futures to finish. Buildbucket handlers always
explicitly wait for futures to finish, so it is not needed.

ndb.toplevel expects all futures to be done by the time
the decorated function call is returned. However, gae_ts_mon
decorator creates a new thread to flush metrics. That thread uses
ndb which means it creates new tasklets. Since it is a separate thread,
by the time toplevel-decorated functions returns, the flush thread
may not return. This makes toplevel think the app deadlocked and it
causes HTTP 500.

Remove ndb.toplevel usage.

R=tandrii@chromium.org, vadimsh@chromium.org

Bug:
Change-Id: I97fdce801f8a1031240414cb1612955a04878bf5
Reviewed-on: https://chromium-review.googlesource.com/690599
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>