[buildbucket] include annotations in Build
Milo builder page that sends a query to buildbucket, needs to render "Info"
column which is annotation's root step Text. (that's what build page does).
For that it needs an annotation in the buildbucket repsonse.

The annotation were excluded because we might hit the datastore entity size
limit. iannucci@ and nodir@ made estimations based on one of long builds,
compressed it and we were still way below limit. Even if we hit the limit,
pubsub handler will start failing for that particular build and the build
won't be updated. We will be able to apply a fix and eventually the build will
be processed.

Stop excluding annotation from build_run_result and compress result_details
property.

R=iannucci@chromium.org
Bug: 769338
Change-Id: I8261e0544a36e3c62f59da6fd7fdd50002d8f8ef
Reviewed-on: https://chromium-review.googlesource.com/691203
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>