[Findit] HttpInterceptor
This change introduces the concept of an optional interceptor to be bound
to an http client at init time.

If set, the interceptor's OnRequest, OnResponse and OnException methods
will be called and the return value of those methods will be used by
the client as the request, response and exception respectively.

The authenticating mechanism has been refactored to be provided by the
interceptor, via header injection. The client will call the interceptor's
GetAuthenticationHeaders method, and apply them accordingly to the
request.

Note that this is required for an upcoming change where we capture
metrics about http requests made to external services via an interceptor.

R=stgao,chanli,lijeffrey,wylieb,katesonia
BUG=768912

Change-Id: I4ad690148fab3b13e326d925ddc31c3192819f1c
Reviewed-on: https://chromium-review.googlesource.com/692841
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>