Update the builders.pyl docs for how the remote_run fields work.
I didn't actually understand how the remote_run_repository
and remote_run_sync_revision fields were being used or supposed
to work, which partially led to the outage in the bug. This
doc update should clarify things; there is a build-side CL
that is also enforcing the logic and adding tests for it.

R=phajdan.jr@chromium.org, mmoss@chromium.org, nodir@chromium.org
BUG=765343

Change-Id: I79e0908b99a55587ec4f4c3f51a5bcae41d68a72
Reviewed-on: https://chromium-review.googlesource.com/699936
Reviewed-by: Paweł Hajdan Jr. <phajdan.jr@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@chromium.org>