[Predator] Fix out-of-index exception.
If the dep_path is 'build' and the file_path is 'buildspec/a/b.cc', exception will be raised if we try to split 'buildspec/a/b.cc' with 'build' + '/'.
Fix is to make sure, we check if file_path starts with 'build/' instead of 'build'. 

Bug: 770717
Change-Id: I026404849122674a0146aab3fcd7c65a94f23a57
Reviewed-on: https://chromium-review.googlesource.com/703677
Commit-Queue: Sharu Jiang <katesonia@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>