[test-results] send one event per test
Here we send additional events to BigQuery, to a seprate table
chrome-infra-events.flakiness.test-results. We're sending one event per
test, because sending all the tests in one event, like we currently do,
sometimes causes the row to be larger than the BQ limit of 1MB.

We're going to see if this approach is reasonable, or if it blows up the
size of the table.

Either way, sending duplicate events to BQ via test-results is only
temporary. Only one (or potentailly neither) of these approaches will
survive.

Bug:765432
Change-Id: Ideab4ee3ec8afd1d7b0bc8ddb5cc0cd5f0e74240
Reviewed-on: https://chromium-review.googlesource.com/707721
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Katie Thomas <katthomas@google.com>