[Findit] Only use logging.exception for non-retriable exceptions.
The interceptor currently uses logging.exception() even for exceptions
that will be retried (Such as ConnectionClosedException) and this
generates noise in the GCP Errors page.

Instead this change makes it so that we log a warning when we know we
will retry the call.

R=stgao,chanli
BUG=768912

Change-Id: I5f1809ad898fa15c00d1837fb983f6a9e5e9e3f3
Reviewed-on: https://chromium-review.googlesource.com/707800
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>