[Predator] Decrease the maximum stack_trace size.
For some entities, even though every property doesn't exceed the maximum size, but it is still possible that the entity is too big. The is probably because the stack_trace is too big(even when it is less than 1048487 bytes).

So we can decrease the maximum stack_trace size that we think it's ok to put to datastore.

Bug: 772110
Change-Id: Ibc382dbad21a5ea90d7f1e73a8bc00744dca66be
Reviewed-on: https://chromium-review.googlesource.com/704001
Commit-Queue: Sharu Jiang <katesonia@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>