[dockerbuild] Enable aarch64, fix "scan-build".
Enable multi-platform packages, notably "arm64/aarch64". This allows the
generation of wheel filenames that specify multiple platforms. Since
arm64 == aarch64, but "pip" doesn't know this, this is a useful thing to
do.

Also fix "scan-build" build instructions. The package builds
anomalously, so we need to override the expected package name. We do
this by having the output wheel be scanned for instead of pre-ordained.
We will correct the name when we stage it for packaging.

BUG=chromium:768318
TEST=None

Change-Id: I2a35a95a31b67c2a07576feb46967cb4507f5c8d
Reviewed-on: https://chromium-review.googlesource.com/709694
Commit-Queue: Daniel Jacques <dnj@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>