[Monorail] Extract a new base class for MonorailRequest.
The motivation for this change is that I want to implement CreateIssue in
WorkEnv and have that used by the web UI, v1 API, and also the current
inbound email interface.  However, inbound email does not have a full
MonorailRequest object or an HTTP request to parse, so I need an object
with a subset of just the auth and perm info that is relevant.

In this CL:
+ Add a new MonorailRequestBase base class for the current two kinds of request
  wrappers that we use for the UI and v1 API code, and extract a little bit of
  common code into that new base class.
+ Make the profiler object part of the request wrapper.  This simplifies several
  method argument lists throughout the app.

R=jeffcarp@chromium.org

Bug: monorail:2985
Change-Id: I518ba865f8be75831bcbf4e28266f67f2cc6ecf1
Reviewed-on: https://chromium-review.googlesource.com/717569
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Jeff Carpenter <jeffcarp@chromium.org>