[Findit] Make response headers json serializable.
This is to fix an exception being raised where the response headers in
the case of a non-200 response are failing to be json serialized. It may
be that the headers is not a plan dict, but another mapping.

This change uses the .items() to make the values into a list of pairs
that *should* be json serializable.

R=chanli,wylieb,lijeffrey,stgao

Bug:
Change-Id: Ife84c9f0304cae2b98b115da4655a8163f9f618b
Reviewed-on: https://chromium-review.googlesource.com/720242
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>