kitchen: Get rid of default git credential helper on Windows.
It is not helpful and extremely annoying (pops up modal prompts, and even when
asked to shut up, takes time to startup (.NET FTW) and prints confusing messages
to console).

Unfortunately, it is specified in "system" gitconfig ($(prefix)/etc/gitconfig),
so the only way to get rid of it is to stop picking up system config (there's
an env var for this).

It means we need to transfer all relevant .gitconfig properties from the system
config into the global config (~/.gitconfig). System config is in fact present
only in our Windows git installation, so this CL accidentally affects only
Windows.

R=nodir@chromium.org, iannucci@chromium.org
BUG=756224

Change-Id: Ie27892d0ef5252e1f0ece8e1ebe429517204e510
Reviewed-on: https://chromium-review.googlesource.com/719475
Reviewed-by: Petr Hosek <phosek@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>