Add configuration_id field for sampling profiler and use in identifier
The configuration_id is a single identifer that uniquely represents the
version pairs and collection state, and is used for that purpose in the
pipeline that generates the regression subtrees. This change uses that
field instead of the corresponding individual fields for consistency and
to facilitate correspondence of the CL attributions with the subtree data.

Change-Id: Ie6849a24670ab6fd131fa2c4c27752d33f98db9a
Reviewed-on: https://chromium-review.googlesource.com/722121
Reviewed-by: Sharu Jiang <katesonia@chromium.org>
Commit-Queue: Mike Wittman <wittman@chromium.org>