[bqschemaupdater] squash 3 flags into 1
Flags -project, -dataset and -table represent one thing: location of the table
to update. Standard SQL uses `<project>.<dataset>.<project>` syntax for that.
Squash the flags into one -table of that format.

Do not hardcode chrome-infra-events project as default.
LUCI builds will be in cr-buildbucket project, not chrome-infra-events (but
a unifying view will be in chrome-infra-events), so only CQ will be using
bqschemaupdater for chrome-infra-events. Not hardcoding a special project is
more generic.

Bug:
Change-Id: Ice06f2fb7554633cb7c8a4a51fa7aa432a1cdca4
Reviewed-on: https://chromium-review.googlesource.com/732159
Reviewed-by: Katie Thomas <katthomas@google.com>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>