[Findit] Flake Analyzer - Moving build info functions to shared location
In preparation for creating a pipeline to map a commit position to an
isolated sha, some functionality should first be moved to a shared
location as part of refactoring efforts for Flake Analyzer.

This CL should not introduce any functional changes, but rather serves
to move code around only.

Bug: 779456
Change-Id: I651befeb0f1f0610d83431939025f741fcaccb36
Reviewed-on: https://chromium-review.googlesource.com/737412
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>