Disable WPT auto-import temporarily by skipping the upload step
wpt-import is currently stuck importing a huge CL because of
https://crbug.com/780055. Since the underlying bugs will take a while to
be fixed, the plan is to import manually (and we have to split the CL
into a series of small CLs to work around the bugs), which is tracked at
https://crbug.com/780700.

Disable the auto import until the whole manual process is complete, in
order to prevent wpt-import being unblocked unintentionally and landing
CLs that bring the tree into an inconsistent state.

Will revert this CL afterwards.

Bug: 780700
Change-Id: I6849b5d7ea6f871d140bdf24ffc16c43f86c6dd3
Reviewed-on: https://chromium-review.googlesource.com/751841
Commit-Queue: Robert Ma <robertma@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>