[Monorail] Fix component auto-cc in case where all are removed.
Compute and pass back a boolean to indicate whether the issue changed.
That is better than treating the dictionary of traces as a boolean,
because sometimes that dictionary is empty even in the case when
we are making a change (e.g., of removing an auto-cc).

R=jeffcarp@chromium.org

Bug: monorail:3190
Change-Id: Ie17eb4f8f164efd82af99c977495c5ec8b389059
Reviewed-on: https://chromium-review.googlesource.com/761059
Reviewed-by: Jeff Carpenter <jeffcarp@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>