Edit chopsui name
I may have given the wrong git URL when I registered 'chops-ui' because
stuff isn't working when it should be.

To fix this I need to manually request the bower team to unregister
and register again. and it looks like it takes several weeks for them to
get back to those requests.

So I'm just going to register it again under the name 'chopsui'.

Bug:760265
Change-Id: I3c99d03673766c2a64ef6d6d973bbe41e29eceaf
Reviewed-on: https://chromium-review.googlesource.com/762147
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Joanna Wang <jojwang@google.com>