[Monorail] Update our work-around to urlize() quirks.
Originally, we produced only plain text issue change notifications, and gmail
would autolink URLs when displaying the text.  Then we added the "View issue"
button and that required an HTML message body.  So, escaped and wrapped the
text to make HTML.  But, then the URL were not clickable, so we called urlize().
However, urlize() does not handle pre-escaped text, specifically it considers
"&" to be a valid part of a URL or email address.  So, we worked around that.

The latest version of urlize() seems to allow "&" and then escape it.  But,
when running locally, my version of urlize() does not.   So in
this CL I am adding regex terms to optionally match both "&" and "&amp;".
And, I also added more tests and made them more clear.

R=jojwang@chromium.org

Bug: monorail:3117
Change-Id: I57742c3cb1b23eb97f95870ed98fb05cc1b16f58
Reviewed-on: https://chromium-review.googlesource.com/764529
Reviewed-by: Joanna Wang <jojwang@google.com>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>