[Predator] Store log in datastore for each crash.
Remove the local log.py under analysis/ and create ndb.Ｍodel log.py in app.common.model instead.
To collect logs spreading through all the code, we'd better store those log to datastore (for a crash, the key of its log in datastore is the same identifiers as the CrashAnalysis result for it.)

The Log() entity is created at the very beginning in crash_pipeline.py, and it will be passed all the way down to record logs.

log pass workflow:
------------------
crash_pipeline.py (create Log() if needed and put it into datastore with crash_identifiers as key.)
    -> predator_app.py (predator_for_chromecrash.py, predator_for_clusterfuzz.py)
        -> predator.py (the core analysis api entry)
            -> changelist_classifier.py

The log will finally be added to result to push back to client in following cls.

Bug: 769950
Change-Id: I3244afba5c0276c67ffefaa5e50aee73784dfb02
Reviewed-on: https://chromium-review.googlesource.com/748163
Commit-Queue: Sharu Jiang <katesonia@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>