[Monorail] Allow dashes in domain part of /u/EMAIL urls.
While investigating a problem with one of the service accounts that uses
our API, I found that I could not view the user profile page for it
because there was a dash in part of the domain name.  The request was
generating a 404 without ever reaching our code.

R=jojwang@chromium.org

Change-Id: Id0a1d24ff78555921298e012d3ba47aaca42d72d
Reviewed-on: https://chromium-review.googlesource.com/780393
Reviewed-by: Joanna Wang <jojwang@google.com>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>