[buildbucket] do not cancel tasks
Given a batch of LUCI builds to schedule, buildbucket creates a swarming task
for each of them, but if creation of any of them fails, all tasks are cancelled,
and nothing gets scheduled. This is inconsistent with other parts of
services.add_many_async that strive to not to fail all builds because of one
of them.

This particularly affects CQ which schedules builds in batches. If at least one of
the builds is a LUCI build with a typo in builder name or builder name is undefined,
none of builds is scheduled and entire CQ is stuck.

Do more fine-grained exception handling and do not cancel tasks once created.

Bug:
Change-Id: I7b1ce1472a7bc4473686ac831433e087aa40aad9
Reviewed-on: https://chromium-review.googlesource.com/787004
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>