buildbucket: return build numbers
If a swarming task creation fails, the build number generated for the
build is lost. If the client code retries retries, it can quickly
exhaust a long range of build numbers until the outage/reason is fixed.

Try to return the generated build number in case it ended up not being
used.

R=vadimsh@chromium.org

Bug:

Change-Id: Idb741d68158012de03badc56dd9dcc2b24fc0ba1
Reviewed-on: https://chromium-review.googlesource.com/474058
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>