[Findit] Flake Analyzer - Give filed bugs priority and a status
When Findit files a bug, it doesn't assign a priority. I fixed this so that now
a priority is set. This will fix ths issue where it wouldn't surface correctly
in SoM. Right now the status is Unconfirmed, but this should actually be
available.

Bug: 789250
Change-Id: I0dc23c939c7a156f921e5341b595f12168e8fc8a
Reviewed-on: https://chromium-review.googlesource.com/794375
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Brandon Wylie <wylieb@chromium.org>