[Findit] Make sure not append a try_job result which is already in results.
Before we check if the the last element in try_job result has the same try_job_id with the new result. But I have noticed some cases where elements with the same try_job_id are in the list.
This change checks all element in list to make sure we only update the existing result if the try_job_id could be found.

Bug:
Change-Id: Ic8ac141ee84ab5200d03a40b22e707e6eca80880
Reviewed-on: https://chromium-review.googlesource.com/794843
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>