[bigquery] refactor BigQueryHelper
After a recent version bump, the lock is no loger necessary. At this
point, the class is no longer necessary either, so remove it.

This is a backwards incompatible change. At the point, BQH has only one
client, CQ, which will be updated in a subsequent CL.

Bug:787231
Change-Id: I37d03b3d36fdb904e5885b3d09e7db1a13be7660
Reviewed-on: https://chromium-review.googlesource.com/797250
Commit-Queue: Katie Thomas <katthomas@google.com>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>