[Predator] Publish back result for every crash.
This change makes sure that we start a CrashWrapperPipeline for every requested crash, and no matter whether the a new analysis is needed for this crash, we will Run PublishResultPipeline to publish result (it might just be {'found': 'False'} if we found nothing or we don't support the analysis at all).

This means we are gonna push result for every requested crash for Clusterfuzz/Cracas/Fracas.
The difference among different clients is that, we only include ``log`` field to explain for the result pushed back to clusterfuzz, but not to Cracas/Fracas (This is because that changes to parse and store `log` field in the result are yet to be added in these 2 clients.)


Bug:
Change-Id: I6cb95c8f196e1e62ba16b33325a738c2824a8291
Reviewed-on: https://chromium-review.googlesource.com/756282
Commit-Queue: Sharu Jiang <katesonia@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>