[Findit] Flake Analyzer - Add pipeline to determine isolated sha of commit position
Introducing a new pipeline to determine the isolated sha of a commit position.
If the commit position maps to that of an existing build, get that build's sha and
return it immediately. Else, a flake try job is needed to compile and isolate at
the requested commit/revision.

Bug: 778533
Change-Id: I9d306cc994463992ad506bb3e313ea5d7b42081a
Reviewed-on: https://chromium-review.googlesource.com/738800
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>
Reviewed-by: Brandon Wylie <wylieb@chromium.org>