[Findit] Fix bad logic for flake tryjob cache assignment.
The reason why the previous flake-specific clause was not working is
that it didn't handle the case where no bots were available or no bots
had the cache name already.

With this change we use the existing check to see if there are available
bots having the cache, and otherwise let the call fall through to the
same else clause as non-flake tryjobs.

R=wylieb,lijeffrey
BUG=800218

Change-Id: Ic4331d4f59a4a6d87fd548e88899b398d157dd71
Reviewed-on: https://chromium-review.googlesource.com/856219
Reviewed-by: Brandon Wylie <wylieb@chromium.org>
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>