Rename AnalyzerProgress to FunctionProgress
ProgressResponse includes results for both isolators and analyzers.

We still want to include both types of function, but will likely
want to differentiate them in the Gerrit plugin UI.

This is one change to switch to using the new terminology and
concept of functions including isolators and analyzers.

Bug: 793735, 799838
Change-Id: I12b68b4c2265becf0a1ae60f269b7541c13ef631
Reviewed-on: https://chromium-review.googlesource.com/861097
Reviewed-by: Emma Söderberg <emso@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>