On workflow launched error, log run ID
This is a minor one line change that should only affect
what is logged in an error case. If I understand correctly,
previously it would always log "1" as the WorkflowRun ID,
but we actually wanted to log the AnalyzeRequest ID, since
that is the "run ID".

Change-Id: Ibeca76ef88aa83b7cd7ceffce473ccd01065417a
Reviewed-on: https://chromium-review.googlesource.com/867942
Reviewed-by: Emma Söderberg <emso@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>