Rename AnalyzerRun to FunctionRun
In this CL:
 - Rename track.AnalyzerRun to track.FunctionRun
 - Rename track.AnalyzerRunResult to track.FunctionRunResult

This change doesn't affect the API, but does affect
the datastore schema.

I plan to use the datastore manager to remove AnalyzerRun
and AnalyzerRunResult entities from the datastore after
this change.

This is expected to cause progress requests
for past runs to return empty results.

Bug: 793735
Change-Id: I3e4e07ec5456b9d142b685afc5acbe0fb222eb26
Reviewed-on: https://chromium-review.googlesource.com/865957
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>