Add Tricium comment UUID to robot comment properties
Comment IDs surfaced in PolyGerrit don't reflect the IDs used
when posting robot comments from Tricium; I believe that this
is intentional and the id field is not meant to be provided
when posting new comments.

This CL would add the ID into a field in properties, in order
to surface the Tricium-generated ID to the frontend so that
it's accessible to the feedback button.

Bug: 800232
Change-Id: Ia0df65bf77240177cddff0c415b506c2b8859c1e
Reviewed-on: https://chromium-review.googlesource.com/872070
Reviewed-by: Emma Söderberg <emso@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>