Rename model fields Analyzer(s) to Functions(s)
This is a follow-up to https://crrev.com/c/865957 and was
originally part of that change. Like that change, this
is a breaking change, meaning that requests which involve
looking up functions or getting the function name from
a worker run, including Progress requests, will not
work for past analyze requests.

Bug: 793735
Change-Id: Iafaa3e737506d1a2ca967c7b8e4f97673b6c4578
Reviewed-on: https://chromium-review.googlesource.com/872116
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Emma Söderberg <emso@chromium.org>