[Tricium] Include revision in Gerrit mapping ID
Reason: Users may want to view users may want to view
Tricium run progress for previous patchsets, which would
show for example which analyzers failed or had comments.

It might be confusing if the progress for the latest
patchset is always shown even if one is viewing previous
patchsets.

In order to support this, the revision string needs to
be included in the Gerrit change strings that are mapped
to run IDs.

Bug: 806532
Change-Id: I541eb692bb3ea2bf5f5fc5ccac56e24d55e180f6
Reviewed-on: https://chromium-review.googlesource.com/890461
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>