Use the git ref when constructing GerritDetails
In the Tricium service, when polling Gerrit,
we construct an AnalyzeRequest message which will then be
used make an Analyze pRPC request. In tricium.proto
it suggests that Revision in GerritDetails is a git ref
string, e.g. refs/changes/56/123456/1, whereas currently
we're using a revision SHA1 hash.

This CL changes the string used to the git ref.

This is a follow-up to https://crrev.com/c/890461.

Bug: 806532
Change-Id: I85ad8f24ef98dbbf066783f140be4987326ef791
Reviewed-on: https://chromium-review.googlesource.com/892359
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>