In set review request, use patch set number
This is expected to fix tricium sending comments
to Gerrit, which stopped working at the last
deployment.

In https://crrev.com/c/892359, the Revision field
in GerritDetails in AnalyzeRequest was changed
to be the ref (e.g. refs/changes/78/78/1).

But, this revision string is used as revision-id
in SetReview requests to the Gerrit REST API.

The revision-id used can be a patch set number
or commit ID hash, but not a ref string. This CL
would change it to be a patch set number, since
this can be easily extracted from the ref string.

https://gerrit-review.googlesource.com/Documentation/rest-api-changes.html#revision-id

Bug: 806532
Change-Id: I8984f298773882ea5be9789a263ed960bfeee42d
Reviewed-on: https://chromium-review.googlesource.com/903222
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>