Test results zip handler: Paginate GCS access
Some zip files we access are larger than 32 MB, which is the size
limit for urlfetch responses. This CL makes the code paginate through
the file, 31 MB at a time.

I picked 31 MB because, while testing on test-results-test.appspot.com,
I found it complained at that threshold. I assume that's because it
doesn't have space for HTTP headers in the response, if GCS sends 32 MB
of data.

Bug: 810427
Change-Id: I296020490d3a347b43913311761eff5c5d127d3a
Reviewed-on: https://chromium-review.googlesource.com/909533
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>