[Findit] Flake Analyzer - Allow for triggering swarming tasks to override isolated input sha
This change introduces a new optional field for triggering swarming tasks against specific
isolated inputs, without affecting existing code. If a sha is not not specified, the reference
swarming task's isolated inputs is used by default (existing behavior).

This is not a permanent solution, as triggering, monitoring, and reporting a swarming task's
results are to be combined into a new pipeline. Rather this is is to unblock testability of
the new Flake Analyzer's pipelines.

Bug: 776958
Change-Id: I6d1be8159d694e456fea93fa5ada33bc43f13ace
Reviewed-on: https://chromium-review.googlesource.com/740952
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Brandon Wylie <wylieb@chromium.org>
Commit-Queue: Jeffrey Li <lijeffrey@chromium.org>