[Monorail] Avoid error on unicode enum values.
We used to always provide "short" versions of values to
be displayed so that the short version would fit in a small area of
the page.  Now we serve the whole value string to the browser and use
CSS overflow to display it within the available space.

R=jojwang@chromium.org

Bug: monorail:3287
Change-Id: I430434a468663dedf426d9219fd8c6304ba2bf8e
Reviewed-on: https://chromium-review.googlesource.com/924308
Reviewed-by: Joanna Wang <jojwang@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>