Add a /run page which shows details given a run ID
Purpose:
Most of the errors logged in other handlers will log a run ID, but not much else.
For now, this will help with debugging other errors. In the initial version, the
run page shows basic information about a run, including a link to the Gerrit change.

Also, eventually I think we'd like to add a user-facing page that we can link
to for run details.

Example page:
https://13988-129681c-tainted-qyearsley-dot-tricium-dev.appspot.com/run/6224707376381952

Bug: 808529
Change-Id: I0b043ea5e39eab5e7650383cc3fa3c9d034e36a6
Reviewed-on: https://chromium-review.googlesource.com/912658
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>