[Findit] Use canonical step_name to decide if Findit supports a step or not.
This change will make Findit to check a step's canonical name to decide if the step is supported or not.

The supported value will also be saved in failure info so that Findit doesn't need to check with config repeatedly in different steps of analysis.

Bug: 812122
Change-Id: I347b370491973f2333dcfd61f6e2a116ad88067e
Reviewed-on: https://chromium-review.googlesource.com/920386
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>