Remove the commented-out frontend pages
This removes the template for the page which would list
recent requests; this could be re-added later.

The reason for this is the general principle of not keeping
commented-out things checked-in, and to only keep what is
currently needed.

This moves a function `requests` which was previously a
private function in the frontend package to be a public
function in the common/track package, since it's used
in tests and might be useful later.

This also resolves a TODO I added earlier.

Change-Id: Ifbd237b350d708142f31c726762af2e22f06ca22
Reviewed-on: https://chromium-review.googlesource.com/927359
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>