[cr-audit-commits] Add fields for custom notification.
The purpose of this change is the eventual introduction of custom
notifications in addition to bug filing for certain rules.

Currently the notifier relies on IssueID being 0 for determining if a
notification has already been sent. This will change, as not all
violations will be reported via Monorail Issue.

This is the first step. Once this is deployed, we'll manually backfill
the field for previous notifications, and then we'll introduce a change
to query the commits to notify on using this flag instead of IssueID.

R=stgao,chanli,lijeffrey,wylieb
BUG=810964

Change-Id: Ie9b77557d3b46b6c04b858a1a63142b50dd6fdfd
Reviewed-on: https://chromium-review.googlesource.com/912648
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>