[Findit] Implement GetCommitsInLastHour.
This function will be used by the upcoming cron job to do full builds on
the bots' caches before they are needed. It will be used to determine if
it is a good time to schedule the builds, as it is thought that the
least bad time to use these bots for such tasks is when nobody is
committing code to the repo.

R=chanli,stgao,lijeffrey,wylieb

Bug: 810959
Change-Id: Idddc686c7c5400c428b2c4e4d01aa888efc9d8c2
Reviewed-on: https://chromium-review.googlesource.com/917461
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>