Filter out deleted files when polling Gerrit
GitFileIsolator was failing only when it encountered files that
were deleted in a change. This only happens when an AnalyzeRequest
was created that includes deleted files.

It turns out that Emma had already thought of filtering out deleted
files when creating analyze requests when polling Gerrit -- the
only reason why they weren't filtered out because the wrong constant
was used to indicate deletion.

This CL adds a unit test to cover the filtering behavior and makes
named constants for the possible Gerrit file statuses.

Bug: 806770
Change-Id: If28e22968b383e948d492e673488b450d2be904e
Reviewed-on: https://chromium-review.googlesource.com/935677
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>