gsubtreed: store synthesized commits in subtree repo
The Commit.alter() method does two things: pass through its arguments
to a CommitData.alter() call, and then intern the new CommitData
object in the existing Commit's repo. But in gsubtreed, that's not
what we want to do: we want to intern the new commit in the
subtree repo, not in the origin repo. This change splits the data
alteration and the repository intering into two steps so that they
can each happen in the right place.

Bug: 816733
Change-Id: If7f843f35e3c255f5bec629c847e7fcf7467cc76
Reviewed-on: https://chromium-review.googlesource.com/938635
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Alan Bram <flyboy@chromium.org>
Commit-Queue: Aaron Gable <agable@chromium.org>