[Monorail] Move spam flag down UI to discourage confusing it with the star
We frequently have instances where issues created 3-4 years ago are being
marked as spam. The classifier doesn't re-evaluate old issues, so the cause
must be users mistaking the spam flag for the star button. The database
backs this up - these issues were classified as spam via the spam threshold.
I think we can curtail the number of accidental spam flags by de-emphasizing
the flag in the UI and making it clear what it does.

See a preview on the bug:
https://bugs.chromium.org/p/monorail/issues/detail?id=3440#c5

Bug: monorail:3440
Change-Id: I7de8539daa8e24b130512033099e3d9d175d6983
Reviewed-on: https://chromium-review.googlesource.com/912244
Commit-Queue: Jeff Carpenter <jeffcarp@chromium.org>
Reviewed-by: Jason Robbins <jrobbins@chromium.org>