[Findit] Add function to get supported compile builders by platform.
This is needed by the build-aheads to figure out which caches to
pre-build. This current naive approach assumes that builders are compile
builders if they have "[B|b]uild" in their names, or are under the
"chromium" master, and uses the existing approach based on the name of
the builder to determine its platform.

R=stgao,chanli,lijeffrey,wylieb
BUG=810959

Change-Id: I723d39c7ae63181c3c357d7c7a42cd443db23ce3
Reviewed-on: https://chromium-review.googlesource.com/936784
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>