[Findit] Add function to check the status of the tree.
This function looks at the latest status of the chromium tree, and
returns true if it can determine that the tree is open, false otherwise.

Note that this doesn't have too much in common with the functionality in
handlers/collect_tree_closures and trying to extract the common parts of
the code, may add unnecessary complexity.

R=stgao,chanli,lijeffrey,wylieb
BUG=810959

Change-Id: Ida23e0b79b253b2404f3d6741b7aa10c24fc434f
Reviewed-on: https://chromium-review.googlesource.com/939658
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>