[Findit] New model to track build ahead jobs.
Note that these jobs are not associated to any failure in particular,
hence no master/builder/buildnumer composite key.

The use case is the following:
The upcoming cron job to schedule build aheads will first check for
ongoing build-aheads, and for that, it needs to query the datastore for
those entries that are still in progress, check with buildbot if they
are done, and if they are, update their status and metadata, and update
the cache statistics.

R=stgao,chanli,lijeffrey,wylieb
BUG=810959

Change-Id: I188d3a899883578c002d8b9a13a4a8bfe591c4b4
Reviewed-on: https://chromium-review.googlesource.com/938972
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>