[recipe_bundler] Fully working implementation.
Knows how to generate local and remote CIPD packages and sets
useful refs for them, as well as tagging them with the resolved
revision.

This version also resolves symlink refs and the commit to fetch
ahead of time, and includes the ref in the CIPD package name
(see docstring for scheme). Doing this enables EXTREME PERFORMANCE
in the noop case (useful for triggering this on many repos
at the same time). :)

Bug: 811974
Change-Id: Ib002eb72080481fee7f7833c79f13acba5058b84
Reviewed-on: https://chromium-review.googlesource.com/923558
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>