[Findit] Platforms to build.
This function decides which platforms a run of the build aheads cron job
should schedule for building.

The function uses as inputs the available bots for each platform, the
number of ongoing buildaheads, and the level of activity in the repo.

This is meant to select a subset of [android, mac, unix, windows] and
the cron job will later decide for a given platform which specific cache
within each select platform to trigger a build for.

R=chanli,stgao,lijeffrey,wylieb
BUG=810959

Change-Id: I52ed1faf9a137c29a3b84ca9b2e483b22f80f11a
Reviewed-on: https://chromium-review.googlesource.com/941508
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>