[cr-audit-commits] Missing clause in MatchesRelevantCommit.
We previously added support for the special value "*" for an
AccountRules rule set, that would match all commits regardless of
author/committer. We missed this other function whose intent is to match
the rule set to a RelevantCommit entity (As opposed to a git.Commit
object, returned by gitiles api).

R=stgao,cmasso

Change-Id: I303330c85ce724c2c8ab4497c35842ad97c723e3
Reviewed-on: https://chromium-review.googlesource.com/946671
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>