[Findit] Flake Detection - Issue filing implementation
Implementing bug filing logic for flake detection. If 3
occurrences of cq false rejection happen in 24 hours a bug
will be considered for filing. If one exists, it's checked
and if it's open and has't been updated for 24 hours
then it will be updated. If it's closed and hasn't updated
in 24 hours another bug will be filed. If it's closed and
has been updated in the last 24 hours then a new issue is
filed. These rules reflect the current implementation of
Chromium-try-flakes.

Bug:814948
Change-Id: I1e5ff30cb2463585558cdabb8ab6daca57ae58d6
Reviewed-on: https://chromium-review.googlesource.com/935484
Commit-Queue: Brandon Wylie <wylieb@chromium.org>
Reviewed-by: Jeffrey Li <lijeffrey@chromium.org>