Validate AnalyzeRequests coming from the analyze queue.
AnalyzeRequests going to the RPC endpoint are checked;
they should be checked in the same way if they come through
the analyze queue.

The expected result of this is that CLs with only deleted
files would now cause an error to be logged in the analyze
queue handler.

Related change:
https://crrev.com/c/956606 Don't create analyze requests
for changes with no paths.

Bug: 803971
Change-Id: I711d756ca4d220230305e5cd6e7f898526bf2619
Reviewed-on: https://chromium-review.googlesource.com/950241
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>